Fixes and improvements

+++ mng/templates/finishall.sh
@@ -41,13 +41,24 @@ if [ -r "./docker.cfg" ]; then
     . "./docker.cfg"
 fi
 
-#if [ -n "${COMPOSE_FILE}" ]; then
-#  F="plain.${COMPOSE_FILE}"
-#  rm -f "$F"
-#  "./luncher.sh" config > "$F"
-#  export COMPOSE_FILE="$F"
-#fi
+if [ -n "${COMPOSE_FILE}" ]; then
+  F="plain.${COMPOSE_FILE}"
+  rm -f "$F"
+  "./luncher.sh" config > "$F"
+  export COMPOSE_FILE="$F"
+fi
 
+
+
+#ip link add dummy0 type dummy >/dev/null 2>&1 
+#if [[ $? -eq 0 ]]; then
+#  # clean the dummy0 link
+#  ip link delete dummy0 >/dev/null 2>&1
+#  PRIVILEGED=*
+#else
+#  PRIVILEGED=
+#fi
+		
 if [ -n "$d" ]; then
     echo "Stop/kill/rm FG GUI App: '$d'..."
 
@@ -62,10 +73,30 @@ fi
 ## TODO: FIXME: in reversed order!?
 for d in ${background_services}; do
 
+    if [ ! "x$d" = "xomd_agent" ]; then  
     echo "Stop/kill/rm BG Service: '$d'..."
-    sleep 20
     "./luncher.sh" stop -t 10 "$d"
     "./luncher.sh" kill -s SIGTERM "$d"
     "./luncher.sh" kill -s SIGKILL "$d"
     "./luncher.sh" rm -f "$d"
+    fi
+    
+done
+
+unset COMPOSE_FILE
+
+#! TODO: Make sure that OMD known the most recent states (trigger update?)
+
+for d in ${background_services}; do
+
+    if [ "x$d" = "xomd_agent" ]; then  
+    echo "Stop/kill/rm BG Service: '$d'..."
+    
+    sleep 5
+    "./luncher.sh" stop -t 10 "$d"
+    "./luncher.sh" kill -s SIGTERM "$d"
+    "./luncher.sh" kill -s SIGKILL "$d"
+    "./luncher.sh" rm -f "$d"
+    fi
+    
 done
